+++ js/nwjs/config.js
@@ -1,9 +1,0 @@
-'use strict'
-
-/**
- * Module dependencies
- */
-const Config = require('configstore');
-const pkg = require('../../package');
-
-module.exports = new Config(pkg.name);