+++ CHANGELOG.md
@@ -24,3 +24,9 @@
 
 # 0.0.7
 * Update vscode-chrome-debug-core to 3.8.2
+
+# 0.0.8
+* Add json schema
+* Generate package.json when not exists
+* Add snapshot compile command `NWjs Compile`
+* Add publish command `NWjs Publish`